refactor: メニュー定義を entities/navigation へ移設
複数コンポーネント（sidebar / nav-sheet / nav-menu）から参照される共通ロジックのため、
nav-menu 配下ではなく entities/navigation スライスに配置し直す。
