+++ application/apps/web/src/client/components/ui/layout/nav-sheet/index.tsx
@@ -9,7 +9,7 @@ import {
   SheetTrigger,
 } from '@/client/components/shadcn/sheet'
 import NavMenu from '@/client/components/ui/navigation/nav-menu'
-import { getVisibleMenuItems } from '@/client/components/ui/navigation/nav-menu/menu-items'
+import { getVisibleMenuItems } from '@/client/entities/navigation'
 import { LogoutButton } from '@/client/features/authenticate/logout'
 
 interface Props {
