+++ application/apps/web/src/client/components/ui/layout/sidebar/index.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/client/components/shadcn/sidebar'
 import { AnchorLink } from '@/client/components/ui/navigation/link'
 import NavMenu from '@/client/components/ui/navigation/nav-menu'
-import { getVisibleMenuItems } from '@/client/components/ui/navigation/nav-menu/menu-items'
+import { getVisibleMenuItems } from '@/client/entities/navigation'
 import { SidebarLogoutButton } from '@/client/features/authenticate/logout'
 
 interface Props {
