+++ application/apps/web/src/client/entities/navigation/index.ts
@@ -1,0 +1,2 @@
+export { getVisibleMenuItems, menuItems } from './model/menu-items'
+export type { MenuItem } from './model/menu-items'
