+++ application/apps/web/src/client/entities/navigation/model/menu-items.ts
